--- pulsar-package-management/src/main/java/pkg/management/storage/bk/BKStorageConfig.java
@@ -1,37 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package pkg.management.storage.bk;
-
-import java.net.URI;
-import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
-import pkg.management.PkgStorageConfig;
-
-@Data
-@Setter
-@Getter
-public class BKStorageConfig implements PkgStorageConfig {
-    URI url;
-    int numReplicas;
-    String bookkeeperClientAuthenticationPlugin;
-    String bookkeeperClientAuthenticationParametersName;
-    String bookkeeperClientAuthenticationParameters;
-}
