--- pulsar-package-management/src/main/java/pkg/management/storage/bk/BKStorageProvider.java
@@ -1,32 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package pkg.management.storage.bk;
-
-import java.util.concurrent.CompletableFuture;
-import pkg.management.PackageStorageProvider;
-import pkg.management.PkgStorage;
-import pkg.management.PkgStorageConfig;
-
-public class BKStorageProvider implements PackageStorageProvider {
-    @Override
-    public CompletableFuture<PkgStorage> getStorage(PkgStorageConfig config) {
-        return CompletableFuture.completedFuture(new BKStorage(config));
-    }
-}
