+++ pulsar-package-manager/src/main/java/org/apache/pulsar/packages/manager/storage/bk/DLOutputStream.java
@@ -1,0 +1,77 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.apache.pulsar.packages.manager.storage.bk;
+
+import io.netty.buffer.ByteBuf;
+import io.netty.buffer.Unpooled;
+import java.io.IOException;
+import java.io.OutputStream;
+import org.apache.distributedlog.AppendOnlyStreamWriter;
+import org.apache.distributedlog.api.DistributedLogManager;
+
+/**
+ * DistributedLog Output Stream.
+ */
+public class DLOutputStream extends OutputStream {
+
+    private final DistributedLogManager distributedLogManager;
+    private final AppendOnlyStreamWriter writer;
+
+    /**
+     * Construct DistributedLog input stream.
+     *
+     * @param distributedLogManager the Distributed Log Manager to access the stream
+     * @param writer
+     */
+    DLOutputStream(DistributedLogManager distributedLogManager, AppendOnlyStreamWriter writer) {
+        this.distributedLogManager = distributedLogManager;
+        this.writer = writer;
+    }
+
+    @Override
+    public void write(int b) throws IOException {
+        byte[] data = new byte[]{ (byte) b };
+        write(data);
+    }
+
+    @Override
+    public void write(byte[] b, int off, int len) throws IOException {
+        ByteBuf data = Unpooled.wrappedBuffer(b, off, len);
+        write(data.array());
+        data.release();
+    }
+
+    @Override
+    public void write(byte[] b) throws IOException {
+        writer.write(b);
+    }
+
+    @Override
+    public void flush() throws IOException {
+        writer.force(false);
+    }
+
+    @Override
+    public void close() throws IOException {
+        writer.markEndOfStream();
+        writer.close();
+        distributedLogManager.close();
+    }
+}
